Move SynchronizationService to separate file
