+++ lib/main.dart
@@ -11,35 +11,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:provider/provider.dart';
 
-class SynchronizationService {
-  Timer? _timer;
-  final Duration _interval = const Duration(minutes: 10);
-
-  void startSyncTimer() {
-    _scheduleNextTask();
-  }
-
-  void _scheduleNextTask() async {
-    _timer?.cancel();
-    await performSynchronizationTask();
-    _timer = Timer(_interval, () async {
-      _scheduleNextTask();
-    });
-  }
-
-  Future<void> performSynchronizationTask() async {
-    try {
-      await syncBlockchain();
-    } catch (e) {
-      displayNotification(e.toString());
-    }
-  }
-
-  void stopSyncTimer() {
-    _timer?.cancel();
-  }
-}
-
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await RustLib.init();
